+++ mqtt/mqtt-edgehub/src/auth/authorization/edgehub.rs
@@ -233,6 +233,24 @@ pub struct ServiceIdentity {
     auth_chain: Option<String>,
 }
 
+impl ServiceIdentity {
+    pub fn identity(&self) -> String {
+        self.identity.clone()
+    }
+    pub fn auth_chain(&self) -> Option<String> {
+        self.auth_chain.clone()
+    }
+}
+
+impl Clone for ServiceIdentity {
+    fn clone(&self) -> Self {
+        ServiceIdentity {
+            identity: self.identity.clone(),
+            auth_chain: self.auth_chain.clone(),
+        }
+    }
+}
+
 impl fmt::Display for ServiceIdentity {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match &self.auth_chain {
